Objective-C Compatible version of library.
Objective-C init method with no parameters, defaults to the same options as the Swift default parameters.
Objective-C init method which allows user to set parameters - passing nil will default the parameters.  InMemoryCache is a BOOL parameter to use or not use a cache.
Objective-C preview method which returns a dictionary of values on success, and an NSError object on failure which contains a localized error description