Fix composition of image URL and Base URL.
When an image URL is not an absolute URL (ie. protocol://url), we want
to put the base URL in front of it, if it is known.

BUT, if neither the base URL ends in / or the image URL start with /,
the current formatImageURL mashes them together without a slash, leading
to something like https://base.comimages/image.jpg